+++ cernera_me/src/components/pages/contact/ContactPage.module.scss
@@ -63,72 +63,3 @@
 
 }
 
-.email-form {
-    color: inherit;
-    font-weight: 500;
-
-    text-align: left;
-    margin-top: 30px;
-    margin-left: auto;
-    margin-right: auto;
-    width: 70%;
-    @media only screen and (max-width: 900px) {
-        width: 100%;
-    }
-
-    &__text-area { 
-        border-style: solid;
-        border-color: #a8a8a8;
-        border-radius: 0px;
-    }
-
-    &__text-area:focus {
-        border: 2px solid #393939;
-
-        outline: none;
-        -webkit-box-shadow: none;
-        -moz-box-shadow: none;
-        box-shadow: none;
-    }
-
-    .button_slide_right {
-        color: white;
-        background-color: #393939;
-        margin-top: 20px;
-        border: 2px solid #393939;
-        border-radius: 0px;
-        padding: 13px 26px;
-        display: inline-block;
-        font-family: 'Raleway', sans-serif;
-        font-weight: normal;
-        font-size: 14px;
-        letter-spacing: 1px;
-        cursor: pointer;
-        box-shadow: inset 0 0 0 0 white;
-    
-        -webkit-transition: ease-out 0.6s;
-        -moz-transition: ease-out 0.6s;
-        transition: ease-out 0.4s;
-
-        @media only screen and (max-width: 750px) {
-            width: 100%;
-            height: 70px;
-        }
-    
-    
-    }
-    
-    .button_slide_right:hover {
-
-        
-
-        @media only screen and (min-width: 751px) {
-            box-shadow: inset 150px 0 0 0 white;
-            background-color: #f4f4f4;
-            border-color: #393939;
-            color: black;
-        }
-
-    }
-
-}
