+++ cernera_me/src/components/pages/contact/ContactPage.tsx
@@ -2,6 +2,7 @@ import React, {useState} from 'react';
 import { useHistory } from "react-router-dom";
 import { Container, Row, Col, Card, Form, Button } from "react-bootstrap";
 import UserCard from "components/user/UserCard";
+import ContactForm from "components/pages/contact/ContactForm";
 import styles from './ContactPage.module.scss';
 import button_styles from '../../Buttons.module.scss';
 import genStyle from "components/General.module.scss";
@@ -20,25 +21,6 @@ const ContactPage = ({ username, allUsers }: { username: string, allUsers: UserI
 
     const user: UserInfoType | undefined = findUser(allUsers, username);
 
-    let [email, setEmail]: any[] = useState({
-        name: '',
-        email_address: '',
-        subject: '',
-        message: ''
-    });
-
-    let handleChange = (e: any) => {
-        let id = e.target.id;
-        let value = e.target.value;
-        email[id] = value;
-        setEmail(email);
-    }
-
-    let handleClick = (e: any) => {
-        e.preventDefault();
-        sendEmail(email);
-    }
-
     return (
         <>
             {user ?
@@ -48,43 +30,7 @@ const ContactPage = ({ username, allUsers }: { username: string, allUsers: UserI
                         <Row className={styles["contact-page__content__container__content"]}>
                             <Col md={12} className={[genStyle["horizontal-center"]].join(' ')}>
                                 <UserCard user={user} />
-                                
-
-                                    <Form className={styles["email-form"]}>
-                                        <Form.Row>
-
-                                            <Col>
-                                                <Form.Group controlId="exampleForm.ControlInput1">
-                                                    <Form.Label>Your name</Form.Label>
-                                                    <Form.Control className={styles["email-form__text-area"]} id="name" placeholder="" onChange={handleChange}/>
-                                                </Form.Group>
-                                            </Col>
-
-                                            <Col>
-                                                <Form.Group controlId="exampleForm.ControlInput2">
-                                                    <Form.Label>Your email</Form.Label>
-                                                    <Form.Control className={styles["email-form__text-area"]} id="email_address" placeholder="" onChange={handleChange}/>
-                                                </Form.Group>
-                                            </Col>
-
-                                        </Form.Row>
-                                        
-                                        <Form.Group controlId="exampleForm.ControlInput3">
-                                            <Form.Label>Subject</Form.Label>
-                                            <Form.Control className={styles["email-form__text-area"]} id="subject" placeholder="" onChange={handleChange}/>
-                                        </Form.Group>
-
-                                        <Form.Group controlId="exampleForm.ControlTextarea1">
-                                            <Form.Label>Message</Form.Label>
-                                            <Form.Control className={styles["email-form__text-area"]} id="message" as="textarea" rows="3" onChange={handleChange}/>
-                                        </Form.Group>
-
-                                        <Button className={styles["button_slide_right"]} variant="primary" type="submit" onClick={handleClick}>
-                                            Submit
-                                        </Button>
-
-                                    </Form>
-
+                                <ContactForm userEmailAddress={user.contact.personal_email}/>
                             </Col>
                         </Row>
                     </Container>
